FrontController 개발

+++ microservices/review-service/src/main/java/se/magnus/microservices/core/review/services/ReviewServiceImpl.java
@@ -2,8 +2,11 @@ package se.magnus.microservices.core.review.services;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.reactivestreams.Publisher;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 import se.magnus.api.core.review.Review;
 import se.magnus.api.core.review.ReviewService;
 import se.magnus.microservices.core.review.persistence.ReviewEntity;
@@ -12,6 +15,9 @@ import se.magnus.util.exceptions.InvalidInputException;
 import se.magnus.util.http.ServiceUtil;
 
 import java.util.List;
+import java.util.function.Supplier;
+
+import static java.util.logging.Level.FINE;
 
 @Slf4j
 @RestController
@@ -21,6 +27,7 @@ public class ReviewServiceImpl implements ReviewService {
     private final ReviewRepository repository;
     private final ReviewMapper mapper;
     private final ServiceUtil serviceUtil;
+    private final Scheduler scheduler;
 
     @Override
     public Review createReview(Review body) {
@@ -36,11 +43,16 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public List<Review> getReviews(int productId) {
+    public Flux<Review> getReviews(int productId) {
         if (productId < 1) {
             throw new InvalidInputException("Invalid productId: " + productId);
         }
 
+        log.info("Will get reviews for product with id={}", productId);
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId)).log(null, FINE));
+    }
+
+    protected List<Review> getByProductId(int productId) {
         List<ReviewEntity> entityList = repository.findByProductId(productId);
         List<Review> list = mapper.entityListToApiList(entityList);
         list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
@@ -51,7 +63,15 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReviews(int productId) {
+        if (productId < 1) {
+            throw new InvalidInputException("Invalid productId: " + productId);
+        }
+
         log.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
         repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
